Fix wrong time interval calculation in Client.query
Network delays can be asymmetric, thus the response from the server can
correspond to any point in time between baseTime and replyTime. If it
does correspond to replyTime (ie. network latency for the request is
large, while network latency for the response is negligible), baseTime
actually corresponds to a time that's earlier than the beginning of the
interval returned by the server. If it corresponds to baseTime, baseTime
corresponds to the interval returned by the server. Thus the interval
returned from the server has to be broadened on the lower side only.

Change-Id: I6fab289a5daa8e3390f426172fa2b084b7c365fb